requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.110.0
openai==1.3.7
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
feedgen==0.9.0
httpx==0.25.2
python-dotenv==1.0.0